Reorganize app directories
