+++ App.tsx
@@ -1,24 +1,11 @@
-import { StyleSheet, SafeAreaView, StatusBar } from 'react-native';
-import { StatusBar as ExpoStatusBar } from 'expo-status-bar';
-import { RecipeList } from './src/components/RecipeList';
+import { StatusBar } from 'expo-status-bar';
+import { RecipesScreen } from './src/screens/RecipesScreen';
 
 const App = (): JSX.Element => (
   <>
-    <SafeAreaView style={styles.container}>
-      <RecipeList />
-    </SafeAreaView>
-    <ExpoStatusBar style="auto" />
+    <RecipesScreen />
+    <StatusBar style="auto" />
   </>
 );
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    flexGrow: 1,
-    marginTop: StatusBar.currentHeight,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
 
 export default App;
