+++ src/utils/layout.ts
@@ -10,3 +10,4 @@ export const SPACING = {
 };
 
 export const WIDTH = Dimensions.get('window').width;
+export const HEIGHT = Dimensions.get('window').height;
